hosts/nixjoy: Implemented Sunshine Dynamic Scaling
Implemented profiles for each device to avoid 4K processing when target
device is a lower resolution. For each profile, nixjoy's display
resolution is changed to match the client's resolution and refresh rate
before starting streaming. For the Framework laptop, we use 2048x1152.

The SUNSHINE_CLIENT_WIDTH, SUNSHINE_CLIENT_HEIGHT, and
SUNSHINE_CLIENT_FPS, does not work well for this because Framework 13
laptop has a non-standard resolution that is not supported by desktop
monitor and monitor modes expect non-integer refresh rates.